requirements.txt
fastapi
passlib
pydantic[email]
pytest
SQLAlchemy
starlette
itsdangerous
tenacity
colorlog
pyjwt[crypto]
httpx
uvicorn
python-multipart
langchain
langchain-google-genai
python-dotenv
jinja2
pytz